% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDescriptionPar.R
\name{addDescriptionPar}
\alias{addDescriptionPar}
\title{addDescriptionPar}
\usage{
addDescriptionPar(target_files)
}
\arguments{
\item{target_files}{Files which should be edited}
}
\value{
Modded text
}
\description{
addDescriptionPar
}
\examples{
\dontrun{
addDescriptionPar()
}
}
\author{
Abhijeet Mishra
}
