% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDescriptionPost.R
\name{addDescriptionPost}
\alias{addDescriptionPost}
\title{addDescriptionPost}
\usage{
addDescriptionPost(target_files)
}
\arguments{
\item{target_files}{Files which should be edited}
}
\value{
Modded text
}
\description{
addDescriptionPost
}
\examples{
\dontrun{
addDescriptionPost()
}
}
\author{
Abhijeet Mishra
}
