% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHeadR.R
\name{addHeadR}
\alias{addHeadR}
\title{addHeadR}
\usage{
addHeadR(model_path = NULL)
}
\arguments{
\item{model_path}{Path to file for cleanup of R block}
}
\value{
Adds header metadata
}
\description{
addHeadR
}
\examples{
\dontrun{
addHeadR()
}
}
\author{
Abhijeet Mishra
}
