% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewlineR.R
\name{addNewlineR}
\alias{addNewlineR}
\title{addNewlineR}
\usage{
addNewlineR(model_path = NULL)
}
\arguments{
\item{model_path}{Path to model}
}
\value{
Modified GMS files with "optimized" parameter descriptions
}
\description{
addNewlineR
}
\examples{
\dontrun{
addNewlineR()
}
}
\author{
Abhijeet Mishra
}
