% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOptimizedPars.R
\name{addOptimizedPars}
\alias{addOptimizedPars}
\title{addOptimizedPars}
\usage{
addOptimizedPars(file_path, return_df = FALSE, post = FALSE)
}
\arguments{
\item{file_path}{Path of files to be modified}

\item{return_df}{If modded dataframe is returned with intermediate edits}

\item{post}{if only clean text is returned for postsolve files}
}
\value{
Modded text
}
\description{
addOptimizedPars
}
\examples{
\dontrun{
addOptimizedPars()
}
}
\author{
Abhijeet Mishra
}
