% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRealizationPhase.R
\name{addRealizationPhase}
\alias{addRealizationPhase}
\title{addRealizationPhase}
\usage{
addRealizationPhase(model_path = NULL)
}
\arguments{
\item{model_path}{Path of files to be modified}
}
\value{
Modded text
}
\description{
addRealizationPhase
}
\examples{
\dontrun{
addRealizationPhase()
}
}
\author{
Abhijeet Mishra
}
