% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanR.R
\name{cleanR}
\alias{cleanR}
\title{CleanR}
\usage{
cleanR(file_path)
}
\value{
Cleanup for GMS files with "optimized" parameter descriptions
}
\description{
CleanR
}
\examples{
\dontrun{
cleanR()
}
}
\author{
Abhijeet Mishra
}
