% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Computes statistical tests of each feature in the dataframe against the dependent variable.}
\usage{
compute(df, dv, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}

\item{depvar}{column name in dataframe of interest}
}
\value{
A flextable of statistical tests
}
\description{
Computes statistical tests of each feature in the dataframe against the dependent variable.
}
\examples{
\dontrun{compute(iris,"Species",c(0,0,0,0,1))}

}
