% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexin.R
\name{flexin}
\alias{flexin}
\title{Return a summary table of a dataframe}
\usage{
flexin(df, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}
}
\value{
flextable
}
\description{
Return a summary table of a dataframe
}
\examples{
\dontrun{flexin(iris,c(0,0,0,0,1))}

}
