% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configs.R
\name{get_configs}
\alias{get_configs}
\title{Retreive configurations for each data type in the GtR API.}
\usage{
get_configs(resource)
}
\arguments{
\item{resource}{Specify the type of data for which you want to retrieve the configuration. Names must be as per \code{gtr_endpoints}.}
}
\value{
A list holding two dataframes - one ('parameters') covering query parameters and another ('fields') holding details about the fields you return.
}
\description{
Allows youto retrieve details of the data configuration for each type of entity (eg, project, organisation) in the GtR API. This information could be useful in helping you structure queries.
}
