% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Function to gather data about linked entities for a specified resource.}
\usage{
get_links(resource, resource_id)
}
\arguments{
\item{resource}{The name of the resource you would like to query/return, as per \code{gtr_endpoints}.}

\item{resource_id}{The ID of the resource.}
}
\value{
A dataframe holding details of any entities linked to the one you specified.
}
\description{
Sends a query to return all records for your chosen resource, with the option to query a combination of resources.
}
