% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_collaboration_outcome_function.R
\name{gtr_collaboration_outcome}
\alias{gtr_collaboration_outcome}
\title{gtr_collaboration_outcome}
\usage{
gtr_collaboration_outcome(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and collaborating organisation output
}
\description{
This function extracts collaborator ouytput data for a project
}
