% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_combinations.R
\docType{data}
\name{gtr_combinations}
\alias{gtr_combinations}
\title{List of endpoints combinations for query_resource_linked_res function.}
\format{
An object of class \code{list} of length 11.
}
\usage{
gtr_combinations
}
\description{
A list of endpoint combinations for the query_resource_linked_res function.
}
\keyword{datasets}
