% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_data_search_function.R
\name{gtr_data_search}
\alias{gtr_data_search}
\title{gtr_data_search}
\usage{
gtr_data_search(dataframe, search_term)
}
\arguments{
\item{dataframe}{Dataframe created using gtr_project_extract_function}

\item{search_term}{vector of search terms to reduce dataframe created using gtr_project_extract_function}
}
\value{
Dataframe with projects from search term - the filtered dataframe
}
\description{
This function seacrches in the title and abstract of a dataframe to filter project data.
}
