% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_endpoints.R
\docType{data}
\name{gtr_endpoints}
\alias{gtr_endpoints}
\title{List of endpoints for the GtR API.}
\format{
An object of class \code{list} of length 18.
}
\usage{
gtr_endpoints
}
\description{
A list of endpoints for the GtR API, used in other parts of the package.
}
\keyword{datasets}
