% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_funder_year_function.R
\name{gtr_funder_year}
\alias{gtr_funder_year}
\title{gtr_funder_year}
\usage{
gtr_funder_year(dataframe, start_year, end_year)
}
\arguments{
\item{dataframe}{dataframe from gtr_project_extract function}

\item{start_year}{start year}

\item{end_year}{end year}
}
\value{
DF datafrmae
}
\description{
This function creates dataframe of funder-year that can be used to create a plot
}
