% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_further_funding_function.R
\name{gtr_further_funding}
\alias{gtr_further_funding}
\title{gtr_further_funding}
\usage{
gtr_further_funding(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Further funding dataframe
}
\description{
This function checks whether a project receives further fudning
}
