% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_further_funding_check_function.R,
%   R/gtr_further_funding_function.R
\name{gtr_further_funding_check}
\alias{gtr_further_funding_check}
\title{gtr_further_funding_check}
\usage{
gtr_further_funding_check(url)

gtr_further_funding_check(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
TRUE/FALSE follow up funding

TRUE/FALSE for follow up funding
}
\description{
This function checks whether a project receives further funding

This function checks whether a project receives further fudning
}
