% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_graph_one_mode_function.R
\name{gtr_graph_one_mode}
\alias{gtr_graph_one_mode}
\title{gtr_graph one-mode}
\usage{
gtr_graph_one_mode(dataframe, weighted, start_year, end_year)
}
\arguments{
\item{dataframe}{dataframe from gtr_project_extract function}

\item{weighted}{TRUE - uses project value, FALSE - uses number of projects}

\item{start_year}{start year}

\item{end_year}{end year}
}
\value{
gs igraph object
}
\description{
This function create a inter-organisational collaboration network (projection from org-project data)
}
