% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_ip_summary_function.R
\name{gtr_ip_summary}
\alias{gtr_ip_summary}
\title{gtr_ip_summary}
\usage{
gtr_ip_summary(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and number of IP output
}
\description{
This function give number of IP outputs per category for a project
}
