% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_policy_function.R
\name{gtr_policy}
\alias{gtr_policy}
\title{gtr_policy}
\usage{
gtr_policy(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and policy outcome
}
\description{
This function provides the policy influence items for a project
}
