% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_policy_geo_function.R
\name{gtr_policy_geo}
\alias{gtr_policy_geo}
\title{gtr_policy_geo}
\usage{
gtr_policy_geo(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project id and number of  regions
}
\description{
This function provides a count of policy influence geography for a project
}
