% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_policy_type_function.R
\name{gtr_policy_type}
\alias{gtr_policy_type}
\title{gtr_policy_type}
\usage{
gtr_policy_type(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project id and number of  types
}
\description{
This function provides a count of policy influence types for a project
}
