% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_potential_impact_function.R
\name{gtr_potential_impact}
\alias{gtr_potential_impact}
\title{gtr_potential_impact}
\usage{
gtr_potential_impact(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe of potential project impact
}
\description{
This function extracts potential project impact statement
}
