% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_extract_function.R
\name{gtr_project_extract}
\alias{gtr_project_extract}
\title{gtr_project_extract}
\usage{
gtr_project_extract(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and all organisaitonal collaborators
}
\description{
This function extracts project information, including organisational collaborators
}
