% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_impact_function.R
\name{gtr_project_impact}
\alias{gtr_project_impact}
\title{gtr_project_impact}
\usage{
gtr_project_impact(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe of project impact
}
\description{
This function extracts project impact information
}
