% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_info_function.R
\name{gtr_project_info}
\alias{gtr_project_info}
\title{gtr_project_info}
\usage{
gtr_project_info(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project information
}
\description{
This function extracts basic project information only
}
