% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_person_function.R
\name{gtr_project_person}
\alias{gtr_project_person}
\title{gtr_project_person}
\usage{
gtr_project_person(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and all individual collaborators
}
\description{
This function extracts project information, including individual collaborators
}
