% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_pub_function.R
\name{gtr_project_pub}
\alias{gtr_project_pub}
\title{gtr_project_pub}
\usage{
gtr_project_pub(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with publication
}
\description{
This function extracts pulciation output for a project
}
