% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_project_search_function.R
\name{gtr_project_search}
\alias{gtr_project_search}
\title{gtr_project_search}
\usage{
gtr_project_search(search_term)
}
\arguments{
\item{search_term}{the search term for projects}
}
\value{
Dataframe with projects from search term
}
\description{
This function searches for projects in GtR associated with the search term
}
