% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_pub_df_function.R
\name{gtr_pub_df}
\alias{gtr_pub_df}
\title{gtr_pub_df}
\usage{
gtr_pub_df(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project_id, journal title, year, number of publications in that journal for a particular year
}
\description{
This function extracts publication information, number of journal articles published in a particular journal per year
}
