% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_pub_perform_function.R
\name{gtr_pub_perform}
\alias{gtr_pub_perform}
\title{gtr_pub_perform}
\usage{
gtr_pub_perform(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project_id,average SJR, average h-index
}
\description{
This function extracts average SJR & H index scores for journal output for a project
}
