% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_spin_out_function.R
\name{gtr_spin_out}
\alias{gtr_spin_out}
\title{gtr_spin_out}
\usage{
gtr_spin_out(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with project id and whether it has a spin out
}
\description{
This function returns a 1 if a project has a spin out and 0 otherwise
}
