% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_studentship_person_extract_function.R
\name{gtr_studentship_person_extract}
\alias{gtr_studentship_person_extract}
\title{gtr_studentship_person_extract}
\usage{
gtr_studentship_person_extract(studentship_id)
}
\arguments{
\item{studentship_id}{project id for studentship from gtr.ukri.org}
}
\value{
Dataframe with studentships and student supervisor details
}
\description{
This function extracts studentship information, student and supervisor name
}
