% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_topic_function.R
\name{gtr_topic}
\alias{gtr_topic}
\title{gtr_topic}
\usage{
gtr_topic(project_id)
}
\arguments{
\item{project_id}{projectid}
}
\value{
Dataframe with project id and topic
}
\description{
This function extracts topic classification for a project (selecting max percentage for multiple topics)
}
