% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_topic_single_function.R
\name{gtr_topic_single}
\alias{gtr_topic_single}
\title{gtr_topic_single}
\usage{
gtr_topic_single(project_id)
}
\arguments{
\item{project_id}{projectid}
}
\value{
topic name
}
\description{
This function extracts a single topic classification for a project (selecting max percentage for multiple topics)
}
