% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertR.R
\name{insertR}
\alias{insertR}
\title{insertR}
\usage{
insertR(model_path = NULL)
}
\arguments{
\item{model_path}{Path to model}
}
\value{
Modified GMS files with "optimized" parameter descriptions
}
\description{
insertR
}
\examples{
\dontrun{
insertR()
}
}
\author{
Abhijeet Mishra
}
