% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscplot.R
\name{miscplots}
\alias{miscplots}
\title{Returns plots for feature selection.}
\usage{
miscplots(df, depvar, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{depvar}{column name in dataframe of interest}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}
}
\value{
A pca plot, randomforest importance plot, and an entropy plot.
}
\description{
Returns plots for feature selection.
}
\examples{
\dontrun{miscplots(iris,"Species",c(0,0,0,0,1))}

}
