% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneplot.R
\name{oneplot}
\alias{oneplot}
\title{Return a univariate plot of each feature in a dataframe.}
\usage{
oneplot(df, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}
}
\value{
A ggplot of every feature in the dataframe.
}
\description{
Return a univariate plot of each feature in a dataframe.
}
\examples{
\dontrun{oneplot(iris,c(0,0,0,0,1))}

}
