% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parGen.R
\name{parGen}
\alias{parGen}
\title{parGen}
\usage{
parGen(lines_after_pattern)
}
\arguments{
\item{lines_after_pattern}{Lines for which parameters need to be generated}
}
\value{
Modded text
}
\description{
parGen
}
\examples{
\dontrun{
parGen()
}
}
\author{
Abhijeet Mishra
}
