% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_lines_from_end.R
\name{remove_empty_lines_from_end}
\alias{remove_empty_lines_from_end}
\title{remove_empty_lines_from_end}
\usage{
remove_empty_lines_from_end(file_path)
}
\arguments{
\item{file_path}{File paths}
}
\value{
Removing empty lines
}
\description{
remove_empty_lines_from_end
}
\examples{
\dontrun{
remove_empty_lines_from_end()
}
}
\author{
Abhijeet Mishra
}
