% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyR.R
\name{removeEmptyR}
\alias{removeEmptyR}
\title{removeEmptyR}
\usage{
removeEmptyR(model_path = NULL, multiline_fix = FALSE)
}
\arguments{
\item{model_path}{Path to model}

\item{multiline_fix}{If multiple newlines should be cleaned. Default is FALSE}
}
\value{
Modified GMS files with "optimized" parameter descriptions
}
\description{
removeEmptyR
}
\examples{
\dontrun{
removeEmptyR()
}
}
\author{
Abhijeet Mishra
}
