% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_multiple_empty_lines.R
\name{replace_multiple_empty_lines}
\alias{replace_multiple_empty_lines}
\title{replace_multiple_empty_lines}
\usage{
replace_multiple_empty_lines(file_path)
}
\arguments{
\item{file_path}{File paths}
}
\value{
Removing empty lines
}
\description{
replace_multiple_empty_lines
}
\examples{
\dontrun{
replace_multiple_empty_lines()
}
}
\author{
Abhijeet Mishra
}
