% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformall.R
\name{transformall}
\alias{transformall}
\title{Plot transformation of all the features in a dataframe.}
\usage{
transformall(df, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}
}
\value{
6 Plots for each feature: original,qqplot,log,sqrt,impute, and standardized.
}
\description{
Plot transformation of all the features in a dataframe.
}
\examples{
\dontrun{transformall(iris,c(0,0,0,0,1))}

}
