% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoplot.R
\name{twoplot}
\alias{twoplot}
\title{Bivariate plots for each feature in the dataframe.}
\usage{
twoplot(df, depvar, catvec)
}
\arguments{
\item{df}{tabular dataframe}

\item{depvar}{column name in dataframe of interest}

\item{catvec}{boolean vector of length ncol(df). 1s for categorical features.}
}
\value{
Patchwork ggplot of each feature plotted against the dependent variable.
}
\description{
Bivariate plots for each feature in the dataframe.
}
\examples{
\dontrun{twoplot(iris,"Species",c(0,0,0,0,1))}

}
