% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDataOutput.R
\name{updateDataOutput}
\alias{updateDataOutput}
\title{updateDataOutput}
\usage{
updateDataOutput(model_path = NULL)
}
\arguments{
\item{model_path}{Path to model}
}
\value{
Modified GMS files with "optimized" parameter descriptions
}
\description{
updateDataOutput
}
\examples{
\dontrun{
updateDataOutput()
}
}
\author{
Abhijeet Mishra
}
