% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_projection_function.R
\name{weighted_projection}
\alias{weighted_projection}
\title{weighted_projection}
\usage{
weighted_projection(DF)
}
\arguments{
\item{DF}{A dataframe with three columns - actor, event, weight}
}
\value{
Dataframe edgelist of the one-mode network with 3 columns - node1, node2, weight
}
\description{
This takes the one-mode projection of a weighted two-mode network. Where the ties in the one-mode network are an aggregation of shared event weights.
}
