% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtr_ip_function.R, R/gtr_ip_summary_function.R
\name{gtr_ip}
\alias{gtr_ip}
\title{gtr_ip}
\usage{
gtr_ip(url)

gtr_ip(url)
}
\arguments{
\item{url}{project URL from gtr.ukri.org}
}
\value{
Dataframe with projects and all organisaitonal collaborators

Dataframe with projects and number of IP output
}
\description{
This function extracts IP project information

This function give number of IP outputs per category for a project
}
