% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmptyR.R
\name{removeEmptyR}
\alias{removeEmptyR}
\title{removeEmptyR}
\usage{
removeEmptyR(model_path = NULL)
}
\value{
Modified GMS files with "optimized" parameter descriptions
}
\description{
removeEmptyR
}
\examples{
\dontrun{
removeEmptyR()
}
}
\author{
Abhijeet Mishra
}
